+++ x/sequencers/keeper/genesis.go
@@ -1,16 +1,15 @@
-package sequencers
+package keeper
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	abci "github.com/tendermint/tendermint/abci/types"
 
-	"github.com/dymensionxyz/dymension-rdk/x/sequencers/keeper"
 	"github.com/dymensionxyz/dymension-rdk/x/sequencers/types"
 )
 
-// InitGenesis initializes the capability module's state from a provided genesis state.
+// InitGenesis initializes the sequencers module's state from a provided genesis state.
 // We return the for ValidatorUpdate only the sequencers set by dymint
-func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) []abci.ValidatorUpdate {
+func (k *Keeper) InitGenesis(ctx sdk.Context, genState types.GenesisState) []abci.ValidatorUpdate {
 	var updates []abci.ValidatorUpdate
 	k.SetParams(ctx, genState.Params)
 
@@ -56,9 +55,10 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 	return updates
 }
 
-// ExportGenesis returns the capability module's exported genesis.
-func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
+// ExportGenesis returns the sequencers module's exported genesis.
+func (k *Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
 	genesis := types.DefaultGenesis()
 	genesis.Params = k.GetParams(ctx)
+
 	return genesis
 }
