+++ x/sequencers/module.go
@@ -132,12 +132,12 @@ func (am AppModule) InitGenesis(ctx sdk.Context, cdc codec.JSONCodec, gs json.Ra
 	// Initialize global index to index in genesis state
 	var genState types.GenesisState
 	cdc.MustUnmarshalJSON(gs, &genState)
-	return InitGenesis(ctx, am.keeper, genState)
+	return am.keeper.InitGenesis(ctx, genState)
 }
 
 // ExportGenesis returns the capability module's exported genesis state as raw JSON bytes.
 func (am AppModule) ExportGenesis(ctx sdk.Context, cdc codec.JSONCodec) json.RawMessage {
-	genState := ExportGenesis(ctx, am.keeper)
+	genState := am.keeper.ExportGenesis(ctx)
 	return cdc.MustMarshalJSON(genState)
 }
